camera_worker/requirements.txt
ultralytics>=8.0.0
easyocr>=1.7.0
opencv-python>=4.9.0
fastapi>=0.110.0
uvicorn>=0.27.0
apscheduler>=3.10.0
supabase>=2.3.0
python-dotenv>=1.0.0
pillow>=10.0.0
aiofiles>=23.0.0
reportlab>=4.0.0
